Switch to a long-running service thread
Instead of doing all the work in hotplug scripts, create a longrunning
service thread that just waits for hotplug scripts to send it SIGHUP,
checks configuration, and starts/kills recording as necessary.

This avoids the need for file locking.  I thought it'd be simpler in
other ways too; we'll see.