restart the recorder process when it stops
If the recorder process exits, try restarting it if appropriate.

And to make sure we do so promptly, make sure our wait ends when it
exits.

Worst case (a full filesystem or other unrepairable condition), we'll
end up restarting it repeatedly.  The one-second sleep in the filename
choosing logic will at least make sure we don't do so more than once a
second.